Add plumbing for SSL/TLS support
This change abstracts out the networking functions so that the
user can provide an SSL/TLS implementation.

Callback functions replace `writev()`, `send()`, and `recv()` (there
is also a callback for error reporting). The default interface remains
unchanged. If the user wants to create a SSL/TLS connection they first
negotiate the connection and then use the new function
`amqp_set_sockfd_full()` to provide the networking implementation for
their SSL/TLS library. The user may provide an optional pointer to data
that is passed through to the networking functions.

Signed-off-by: Michael Steinert <mike.steinert@gmail.com>