autotools: Don't include man-date.ent as part of dist package
man-date.ent is a file generated by Makefile.am which puts the current
date in the man pages as they are generated.

The filename was listed in EXTRA_DIST which caused failures in
make dist when --disable-docs or --disable-tools was passed to the
configure script.

Fix is to remove man-date.ent as it really shouldn't be a part of
the distribution package