autoconf: remove incorrect --enable-64-bit flag
This is a fix for issue #64: autotools: --enable-64-bit doesn't work
consistently. This flag doesn't work correctly in the presence of other
CFLAGS variables, or if its not specified and being built on a 64-bit
platform the output at the end of the configure script will be incorrect
causing confusion.

The fix is to remove this flag from the configure.ac script and allow
the users to force building a 32-bit or 64-bit library by setting the
CFLAGS/LDFLAGS appropriate for their compiler/platform at ./configure
time.