Adding basic error handling to SSL send/recv funcs
For recv/send() functions, the functions themselves return a negative
number on error, then the amqp_ssl_socket_error() can be used to return
an opaque (e.g., not very useful to API users) error code. Internally we
store this error in the last_error field of the amqp_ssl_socket_t
object.