ssl: OpenSSL should require explicitly un-init
Instead of un-initializing OpenSSL when the connection count reaches 0
require API users to explicitly un-initialize OpenSSL if they care to.
In the common case this will require much less work to start a new SSL
connection.

As a side effect: cleanup and simplify the OpenSSL initialization
routines, including removing the thread-safety checks done for the BIO
since they will be called under the OpenSSL locking.