cmake: change BUILD_EXAMPLES to default to OFF
This makes the project easier to use as an add_subdirectory in a larger
project without needing to explicitly disable this feature.

As a minor side-effect: BUILD_EXAMPLES now requires
BUILD_SHARED_LIBS=ON. While there isn't a technical limitation, it makes
the CMake code easier to reason about, and is believed this will not
impact a lot of users. If this ends up being a problem, it can be
reconsidered.