ssl: use SSL_get1_peer_certificate() in OpenSSL 3
The function SSL_get_peer_certificate() was deprecated in 3.0.0. Now it is defined as a compatibility macro to SSL_get1_peer_certificate().

However if your OpenSSL 3 is not compiled with the deprecated API included, rabbitmq-c will fail to compile. This commit adds an OpenSSL version check to use the appropriated function.